+++ lib/counter_controller.dart
@@ -10,10 +10,18 @@ class CounterController {
 
   void incrementCounter(){
 
-    model.modifierCoounter(1, add);
+    model.modifierCoounter(1, _add);
   }
 
-  int add (int number1, int number2){
+  void subtractCounter(){
+    model.modifierCoounter(1, _sub);
+  }
+
+  int _add (int number1, int number2){
     return number1 + number2;
+  }
+
+  int _sub(int number1, int number2){
+    return number1-number2;
   }
 }
