+++ lib/counter_model.dart
@@ -6,7 +6,7 @@ int get counter {
 }
 
 void modifierCoounter(int valor, Function(int, int) operation){
-  _counter = operation(valor, _counter);
+  _counter = operation(_counter, valor);
 
   }
 }
